+++ src/App.jsx
@@ -1,11 +1,9 @@
 import { Outlet } from "react-router-dom"
-import RegisterContainer from "./containers/RegisterContainer/RegisterContainer"
 
 function App() {
   return (
     <div>
       <Outlet />
-      <RegisterContainer />
     </div>
   )
 }
