+++ src/components/index.js
@@ -1,0 +1,5 @@
+import Register from './register/register.component'
+
+export {
+  Register
+}
