+++ src/containers/index.js
@@ -1,0 +1,5 @@
+import RegisterContainer from './register/register.container'
+
+export {
+  RegisterContainer
+}
