--- src/pages/register/register.jsx
@@ -1,9 +1,0 @@
-function Register() {
-  return (
-    <div>
-      this is register
-    </div>
-  )
-}
-
-export default Register
