+++ src/pages/register/register.page.jsx
@@ -1,0 +1,11 @@
+import { RegisterContainer } from '../../containers'
+
+function Register() {
+  return (
+    <>
+      <RegisterContainer />
+    </>
+  )
+}
+
+export default Register
